requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai-whisper==20231214
sounddevice==0.4.6
soundfile==0.13.1
numpy==2.2.6
torch==2.0.1
torchaudio==2.0.2
